+++ root/branches/will/options/source/photon/plugins/org.marketcetera.photon.marketdata/src/model/java/org/marketcetera/photon/model/marketdata/MDLatestTick.java
@@ -61,6 +61,6 @@ public interface MDLatestTick extends MDItem {
      * @model suppressedSetVisibility="true"
      * @generated
      */
-    Integer getMultiplier();
+    BigDecimal getMultiplier();
 
 } // MDLatestTick
